refactor(bus): split tick orchestration into its own module
Move the per-cycle peripheral tick walk + DMA copy/signal routing + NVIC/
interrupt-matrix aggregation (11 methods + pend_nvic, ~590 lines) out of
bus/mod.rs into bus/tick.rs. Methods called back from mod.rs (incl. the
event-scheduler-gated sync_scheduler_peripheral / collect_scheduled_events /
dma_copy_unit — parent can't see a child's private items) made pub(crate).

bus/mod.rs 3046 -> 2467. Pure move — verified on BOTH feature sets (default +
jit,event-scheduler): workspace lib + firmware_survival (26) + register_coverage,
clippy clean.
